Crate pallet_anchors

Anchors pallet for runtime

This pallet provides functionality of Storing anchors on Chain

Re-exports

pub use pallet::*;

pub use weights::*;

Modules

pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

weights

Autogenerated weights for pallet_anchors

Structs

AnchorData

The data structure for storing committed anchors.

PreCommitData

The data structure for storing pre-committed anchors.

RootHashSize