Function pallet_anchors::pallet::dispatchables::commit

pub fn commit<T: Config>(
    anchor_id_preimage: T::Hash,
    doc_root: T::Hash,
    proof: T::Hash,
    stored_until_date: T::Moment
)

Commits a document_root of a merklized off chain document in Centrifuge p2p network as the latest version id(anchor_id) obtained by hashing anchor_id_preimage. If a pre-commit exists for the obtained anchor_id, hash of pre-committed signing_root + proof must match the given doc_root. Any pre-committed data is automatically removed on a succesful commit and the reserved funds from Pallet::pre_commit() are returned to the same account. To avoid state bloat on chain, the committed anchor would be evicted after the given stored_until_date. The calling account would be charged accordingly for the storage period. For a more detailed explanation refer section 3.4 of Centrifuge Protocol Paper

NOTE: This function is an automatically generated, doc only, uncallable stub. designed to document the commit variant of the pallet Call enum. You should not attempt to call this function directly.