Trait pallet_anchors::pallet::Config

pub trait Config: Config + Config {
    type Fees: Fees<AccountId = Self::AccountId, Balance = <<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::Balance>;
    type CommitAnchorFeeKey: Get<<Self::Fees as Fees>::FeeKey>;
    type PreCommitDepositFeeKey: Get<<Self::Fees as Fees>::FeeKey>;
    type WeightInfo: WeightInfo;
    type Currency: ReservableCurrency<Self::AccountId>;
}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Required Associated Types

type Fees: Fees<AccountId = Self::AccountId, Balance = <<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::Balance>

Entity used to pay fees

type CommitAnchorFeeKey: Get<<Self::Fees as Fees>::FeeKey>

Key used to retrieve the fee balances in the commit method.

type PreCommitDepositFeeKey: Get<<Self::Fees as Fees>::FeeKey>

Key to identify the amount of funds reserved in a Pallet::pre_commit() call. These funds will be unreserved once the user make the [commit()] succesfully or call Pallet::evict_pre_commits()

type WeightInfo: WeightInfo

Type representing the weight of this pallet

type Currency: ReservableCurrency<Self::AccountId>

Currency as viewed from this pallet