Type Alias pallet_anchors::pallet::Module

pub type Module<T> = Pallet<T>;

👎Deprecated: use Pallet instead

Type alias to Pallet, to be used by construct_runtime.

Generated by pallet attribute macro.

Aliased Type

struct Module<T>(/* private fields */);

Implementations

impl<T: Config> Pallet<T>

pub fn pre_commit( origin: OriginFor<T>, anchor_id: T::Hash, signing_root: T::Hash ) -> DispatchResult

Obtains an exclusive lock to make the next update to a certain document version identified by anchor_id on Centrifuge p2p network for a number of blocks given by PRE_COMMIT_EXPIRATION_DURATION_BLOCKS value. signing_root is a child node of the off-chain merkle tree of that document. In Centrifuge protocol, a document is committed only after reaching consensus with the other collaborators on the document. Consensus is reached by getting a cryptographic signature from other parties by sending them the signing_root. To deny the counter-party the free option of publishing its own state commitment upon receiving a request for signature, the node can first publish a pre-commit. Only the pre-committer account in the Centrifuge chain is allowed to commit a corresponding anchor before the pre-commit has expired. Some funds are reserved on a succesful pre-commit call. These funds are returned to the same account after a succesful Pallet::commit() call or explicitely if evicting the pre-commits by calling Pallet::evict_pre_commits(). For a more detailed explanation refer section 3.4 of Centrifuge Protocol Paper

pub fn commit( origin: OriginFor<T>, anchor_id_preimage: T::Hash, doc_root: T::Hash, proof: T::Hash, stored_until_date: T::Moment ) -> DispatchResult

Commits a document_root of a merklized off chain document in Centrifuge p2p network as the latest version id(anchor_id) obtained by hashing anchor_id_preimage. If a pre-commit exists for the obtained anchor_id, hash of pre-committed signing_root + proof must match the given doc_root. Any pre-committed data is automatically removed on a succesful commit and the reserved funds from Pallet::pre_commit() are returned to the same account. To avoid state bloat on chain, the committed anchor would be evicted after the given stored_until_date. The calling account would be charged accordingly for the storage period. For a more detailed explanation refer section 3.4 of Centrifuge Protocol Paper

pub fn evict_pre_commits( origin: OriginFor<T>, anchor_ids: BoundedVec<T::Hash, ConstU32<EVICT_PRE_COMMIT_LIST_SIZE>> ) -> DispatchResult

Initiates eviction of pre-commits that has expired given a list on anchor ids. For each evicted pre-commits, the deposit holded by Pallet::pre_commit() call will be returned to the same account that made it originally.

pub fn evict_anchors(origin: OriginFor<T>) -> DispatchResult

Initiates eviction of expired anchors. Since anchors are stored on a child trie indexed by their eviction date, what this function does is to remove those child tries which has date_represented_by_root < current_date. Additionally it needs to take care of indexes created for accessing anchors, eg: to find an anchor given an id.

impl<T: Config> Pallet<T>

pub fn get_pre_commit<KArg>( k: KArg ) -> Option<PreCommitData<T::Hash, T::AccountId, T::BlockNumber, <<T as Config>::Currency as Currency<<T as Config>::AccountId>>::Balance>>where KArg: EncodeLike<T::Hash>,

PreCommits store the map of anchor Id to the pre-commit, which is a lock on an anchor id to be committed later

impl<T: Config> Pallet<T>

pub fn get_anchor_evict_date<KArg>(k: KArg) -> Option<u32>where KArg: EncodeLike<T::Hash>,

Map to find the eviction date given an anchor id

impl<T: Config> Pallet<T>

pub fn get_anchor_id_by_index<KArg>(k: KArg) -> Option<T::Hash>where KArg: EncodeLike<u64>,

Map to find anchorID by index

impl<T: Config> Pallet<T>

pub fn get_latest_anchor_index() -> Option<u64>

Latest anchored index that was recently used

impl<T: Config> Pallet<T>

pub fn get_latest_evicted_anchor_index() -> Option<u64>

Latest evicted anchor index. This would keep track of the latest evicted anchor index so that we can start the removal of AnchorEvictDates index from that index onwards. Going from AnchorIndexes => AnchorEvictDates

impl<T: Config> Pallet<T>

pub fn get_latest_evicted_date() -> Option<u32>

This is to keep track of the date when a child trie of anchors was evicted last. It is to evict historic anchor data child tries if they weren’t evicted in a timely manner.

impl<T: Config> Pallet<T>

pub fn get_evicted_anchor_root_by_day<KArg>( k: KArg ) -> Option<BoundedVec<u8, RootHashSize<T::Header>>>where KArg: EncodeLike<u32>,

Storage for evicted anchor child trie roots. Anchors with a given expiry/eviction date are stored on-chain in a single child trie. This child trie is removed after the expiry date has passed while its root is stored permanently for proving an existence of an evicted anchor.

impl<T: Config> Pallet<T>

pub fn get_anchor_by_id( anchor_id: T::Hash ) -> Option<AnchorData<T::Hash, T::BlockNumber>>

Get an anchor by its id in the child storage

pub fn anchor_storage_key(storage_key: &[u8]) -> Vec<u8>