Crate pallet_anchors

Anchors pallet for runtime

This pallet provides functionality of Storing anchors on Chain

Re-exports

pub use pallet::*;

pub use weights::*;

Modules

pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

weights

Structs

AnchorData

The data structure for storing committed anchors.

PreCommitData

The data structure for storing pre-committed anchors.

RootHashSize

Constants

PRE_COMMIT_EXPIRATION_DURATION_BLOCKS

Expiration duration in blocks of a pre-commit This is the maximum expected time for document consensus to take place between a pre-commit of an anchor and a commit to be received for the pre-committed anchor. Currently we expect to provide around 80 mins for this. Since our current block time as per chain_spec.rs is 6s, we set this to 80 * 60 secs / 6 secs/block = 800 blocks.